requirements.txt
selenium
pytest
selenium-wire
allure-pytest